scripts/requirements-arxiv.txt
pymupdf
pdfminer.six
pylatexenc
